Restructure of 'Database' abstract class
Adding 'SQLAlchemyDatabase' to subclass
